% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pceg_function.R
\name{ahc_merge}
\alias{ahc_merge}
\title{The AHC Merge Function}
\usage{
ahc_merge(y1, y2, a1, a2, structural_zero)
}
\arguments{
\item{y1}{An integer count vector for the first stage.}

\item{y2}{An integer count vector for the second stage.}

\item{a1}{A numeric prior vector for the first stage}

\item{a2}{A numeric prior vector for the first stage}
}
\value{
A numeric value for the change in log marginal likelihood
}
\description{
Calculates the change in log marginal likelihood from merging two count vectors given their prior vectors
}
\examples{
ahc_merge(c(1,1),c(0,2),c(1,1),c(1,1))
}
