% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_trees_functions.R
\name{background_extractor}
\alias{background_extractor}
\title{Background Extractor}
\usage{
background_extractor(data, mod.background)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if included) are the last two columns.}

\item{mod.background}{A StagedTree object fit on a subset of data, from the first variable to some cutpoint variable.}
}
\value{
A data set where the variables from the background model, besides the cutpoint variable, have been replaced based on the stage structure for the cutpoint variable.
}
\description{
Takes a background model which produces a new categorisation of certain variables, and replaces them in data.
}
\details{
Intended for use with Intermediate CEGs, this function takes a data set and background model which has been fit on a subset of the data in the same order, and replaces the background variables with their categorisation from the background model.
}
\examples{
mod<-pceg(knee_pain_obs[,1:3],2,FALSE,FALSE)
background_extractor(knee_pain_obs,mod)
}
