% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_trees_functions.R
\name{background_imputer}
\alias{background_imputer}
\title{Background Imputer Function}
\usage{
background_imputer(data, background.data)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if included) are the last two columns.}

\item{background.data}{A data set where each column besides the last corresponds to the same column as}
}
\value{
A data set where the variables in background.data are replaced by their corresponding category for a new variable "background", and the rest of the variables in data are unchanged.  of
}
\description{
Imputes a new categorisation of the background variables.
}
\details{
Intended to be used for Intermediate CEGs, this function takes a data set and a definition of new categorisations of certain background variables in the data. This categorisation can be obtained from the [\code{\link[=cat_extractor]{cat_extractor()}}] function.
}
\examples{
backdata<-cat_extractor(mod)
background_imputer(knee_pain_obs, backdata)
}
