% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pceg_function.R
\name{bayes_factor}
\alias{bayes_factor}
\title{The log Bayes factor Between nested models}
\usage{
bayes_factor(sample_sum, prior, stage1, stage2)
}
\arguments{
\item{sample_sum}{A list where each element is an integer sum of the event counts for each stage.}

\item{prior}{A list of numeric vectors where the numeric vectors are the hyperparameters of the Gamma prior distribution for each stage.}

\item{stage1}{An integer specifying the first stage to be merged.}

\item{stage2}{An integer specifying the second stage to be merged.}
}
\value{

}
\description{
This function calculates the change in log marginal likelihood of a model when two stages are merged into one.
}
