% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_resizer.R
\name{binary_resizer}
\alias{binary_resizer}
\title{The Binary Resizer Function}
\usage{
binary_resizer(data, poisson_response = TRUE, variable_time = TRUE)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}
}
\value{
A data set where each variable is binary.
}
\description{
Reparameterise any data set into one where each non-binary variable is replaced by dummy variables to account for the number of categories.
}
\examples{
v1<-factor(sample(c(0,1,2),10000,TRUE))
data<-cbind(v1,knee_pain_obs)
summary(data)
summary(binary_resizer(data))
}
