% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_trees_functions.R
\name{cat_extractor}
\alias{cat_extractor}
\title{Category Extractor Function}
\usage{
cat_extractor(mod)
}
\arguments{
\item{mod}{A StagedTree object.}
}
\value{
A data frame where the last column corresponds to the stage for each path.
}
\description{
Looks at each possible category combination in a StagedTree model, and returns what stage they are in for the final level.
}
\details{
Intended to be used for Intermediate CEGs, this function looks at each path in the StagedTree model up to the final level, and records what stage it is in for this level.This will define the categories for an Intermediate CEG.
}
\examples{
mod<-pceg(knee_pain_obs[,1:3],2,FALSE,FALSE)
cat_extractor(mod)
}
