% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_trees_functions.R
\name{cat_replacer}
\alias{cat_replacer}
\title{Category Replacer}
\usage{
cat_replacer(X, Y)
}
\arguments{
\item{X}{A data set.}

\item{Y}{A data set with one more column than X.}
}
\value{
A data set where the categories in X are placed by the final column in Y.
}
\description{
Replace the categories in X with the corresponding final column in Y
}
