% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm_functions.R
\name{cdll}
\alias{cdll}
\title{The complete data log likelihood for a Poisson response}
\usage{
cdll(params, y, t)
}
\arguments{
\item{params}{The parameters for the distribution}

\item{y}{The observed count vector}

\item{t}{The observed time vector}
}
\value{
A list containing the likelihood value and a vector of weights for the em algorithm
}
\description{
This function calculates the complete data log likelihood for a count vector and time vector, given values of the parameters of a Poisson distribution.
}
