% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions_S3.R
\name{ChainEventGraph}
\alias{ChainEventGraph}
\title{The ChainEventGraph constructor function.}
\usage{
ChainEventGraph(staged.tree)
}
\arguments{
\item{staged.tree}{An object in the S3 class StagedTree}
}
\value{
An object in the S3 class ChainEventGraphs
}
\description{
Create an object in the S3 class ChainEventGraph using an object in the S3 class StagedTree.
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
ceg<-ChainEventGraph(mod)
plot(ceg)
}
