% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{CheckAndCleanData1}
\alias{CheckAndCleanData1}
\title{The Check And Clean Data Function}
\usage{
CheckAndCleanData1(data.frame)
}
\arguments{
\item{data.frame}{A data set}
}
\value{
A cleaned data set
}
\description{
The function \link[ceg:CheckAndCleanData]{ceg::CheckAndCleanData}
}
