% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{CheckForCleanData1}
\alias{CheckForCleanData1}
\title{The Check For Clean Data Function}
\usage{
CheckForCleanData1(data.frame)
}
\arguments{
\item{data.frame}{A data set}
}
\value{
A logical value indicating whether the data set is clean, along with a message if it isn't
}
\description{
The function \link[ceg:CheckForCleanData]{ceg::CheckForCleanData}
}
