% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_tree_functions.R
\name{counter}
\alias{counter}
\title{The Counter Function}
\usage{
counter(x, v)
}
\arguments{
\item{x}{A value}

\item{v}{A vector}
}
\value{
An integer
}
\description{
Counts how many elements of a vector match a specific value
}
