% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm_functions.R
\name{em_zip}
\alias{em_zip}
\title{The EM algorithm}
\usage{
em_zip(
  data,
  p_0 = NULL,
  l_0 = NULL,
  initial_method = NULL,
  variable_time = TRUE,
  max_iter = 10000,
  tol = 1e-10
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if variable) are the last two columns.}

\item{p_0}{A numeric vector of initial values for the risk probabilities. If the vector is of length 1, this value will be repeated for each leaf. If NULL, then an automatic method (initial_method) will be used.}

\item{l_0}{A numeric vector of initial values for the rates. If the vector is of length 1, this value will be repeated for each leaf. If NULL, then an automatic method (initial_method) will be used.}

\item{initial_method}{A character string indicating the method for automatically setting the initial conditions. The character string can be an element of c("mean",mle","mm"), with "mean" being the default when p_0, l_0 are not provided.}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}

\item{max_iter}{An integer for the maximum number of iterations for the algorithm.}

\item{tol}{A numeric which represents the minimum change in the complete data log likelihood needed to continue the algorithm.}
}
\value{
A list containing a matrix with the desired outputs for each evolution of the process, as well as vectors of the estimates for the rates and risk probabilities.
}
\description{
This function carries out the Expectation-Maximisation algorithm across all leaves.
}
\details{
For each leaf, which represents a unique evolution of the process, the \code{\link[=em]{em()}} function is applied. Initial conditions can either be provided through
}
\examples{
em_zip(knee_pain_obs)
}
