% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_new_functions.R
\name{event.tree.creator}
\alias{event.tree.creator}
\title{The Event Tree Creator}
\usage{
event.tree.creator(data, poisson_response = TRUE, variable_time = TRUE)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}
}
\value{
An object of the class "Stratified.event.tree"
}
\description{
A function to create an object of the "Stratified.event.tree" S4 class that is compatible with Poisson response variables.
}
\examples{
tree<-event.tree.creator(knee_pain_obs,TRUE,TRUE)
plot(tree)
}
