% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions_S3.R
\name{EventTree}
\alias{EventTree}
\title{The EventTree constructor function}
\usage{
EventTree(data, poisson_response = TRUE, variable_time = TRUE, zip = FALSE)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{zip}{A logical value indicating whether the process is zero-inflated (TRUE) or not (FALSE). If the process is zero-inflated but this has already been accounted for in the data input, this should still be FALSE.}
}
\value{
An object of the S3 class EventTree
}
\description{
Create an object in the StagedTree S3 class from a data set.
}
\examples{
tree<-EventTree(knee_pain_obs,TRUE,TRUE)
plot(tree)
}
