% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_counts_function.R
\name{expected_count_calculator}
\alias{expected_count_calculator}
\title{The expected count function.}
\usage{
expected_count_calculator(
  data,
  ceg,
  limit = 4,
  poisson_response = TRUE,
  poisson_time_variable = TRUE,
  posterior = TRUE,
  zip = TRUE
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if variable) are the last two columns}

\item{ceg}{A ceg model fit to the data set, as produced by pceg().}

\item{limit}{An integer where the number of event counts greater than or equal to this integer are grouped together.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{poisson_time_variable}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}

\item{posterior}{A logical value indicating whether the estimates of the parameters used should be the posterior estimate (TRUE) or sample estimate (FALSE).}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}
}
\value{
A list of three matrices. The first matrix is the observed count matrix, the second is the expected count matrix, and the third is the chi-squared contribution matrix.
}
\description{
A function to test the goodness of fit of a pceg model to a data set.
}
\details{
For each leaf stage in a pceg model created by \code{\link[=pceg]{pceg()}}, this function calculates the observed event counts up to some integer limit, and then calculates
the expected event counts for that leaf based on the model estimates of the parameters for the Poisson distribution. Then, for each leaf $i$ and each event count $j$,
a chi-square contribution is calculated using the formula $\frac{(O_{ij}-E_{ij})^2}{E_{ij}}$ and output as a matrix.
}
