% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_functons.R
\name{gibbs_zip}
\alias{gibbs_zip}
\title{The Gibbs Sampler}
\usage{
gibbs_zip(data, N = 1000, variable_time = TRUE, a = 1, b = 1, c = 1, d = 1)
}
\arguments{
\item{data}{A data set where the observed count vector and time vector (if variable) are the last two columns.}

\item{N}{An integer specifying the number of iterations for the Gibbs sampler.}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}

\item{a}{A numeric value for the shape hyperparameter of the Gamma prior for the Poisson rate.}

\item{b}{A numeric value for the rate hyperparameter of the Gamma prior for the Poisson rate.}

\item{c}{A numeric value for the alpha hyperparameter of the Beta prior for the risk probability.}

\item{d}{A numeric value for the beta hyperparameter of the Beta prior for the risk probability.}
}
\value{
A list containing a matrix with the desired outputs for each evolution of the process, as well as vectors of the estimates for the rates and risk probabilities.
}
\description{
A function to carry out Gibbs sampling for each unique covariate combination in the data set.
}
