% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_optimisation_functions.R
\name{glogitinv}
\alias{glogitinv}
\title{The logistic function}
\usage{
glogitinv(p)
}
\arguments{
\item{p}{A numeric value on the real line.}
}
\value{
A numeric value between 0 and 1.
}
\description{
This function maps a value on the real line to a value between 0 and 1. It is the inverse of the \code{\link[=glogit]{glogit()}} function.
}
\examples{
glogitinv(-0.4054651)
}
