% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_creators.R
\name{hdi_beta_extractor}
\alias{hdi_beta_extractor}
\title{The Beta Highest Density Interval Creator}
\usage{
hdi_beta_extractor(
  data,
  mod,
  ci = 0.95,
  N = 10000,
  level_rel_final = -1,
  poisson_response = TRUE,
  variable_time = TRUE,
  zip = TRUE
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if variable) are the last two columns}

\item{mod}{A StagedTree model fit to the data set, as produced by \code{\link[=pceg]{pceg()}}.}

\item{ci}{A numeric value between 0 and 1 specifying the perecentage confidence for the highest density interval.}

\item{N}{An integer specifying the number of iterations for the empirical functions used.}

\item{level_rel_final}{A non-positive integer indicating where the desired variable is relative to the response variable.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}
}
\value{
A series of highest density intervals for each parameter of interest.
}
\description{
This function calculates the highest density interval for the parameters of a Dirichlet posterior, using a Beta distribution.
The Dirichlet posterior is present for the situations with a categorical response variable in any type of CEG.
}
\details{
This function calculates the highest density interval for the estimated transition probabilities for the situation stages in a PCEG, ZIPCEG or CEG.
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
hdi_beta_extractor(knee_pain_obs,mod,zip=FALSE)
}
