% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_creators.R
\name{hdi_gamma_extractor}
\alias{hdi_gamma_extractor}
\title{The Gamma Highest Density Interval Creator}
\usage{
hdi_gamma_extractor(
  data,
  ceg,
  ci = 0.95,
  N = 10000,
  variable_time = TRUE,
  zip = TRUE
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if variable) are the last two columns}

\item{ceg}{A ceg model fit to the data set, as produced by CEG.AHC.POISS.}

\item{ci}{A numeric value between 0 and 1 specifying the perecentage confidence for the highest density interval.}

\item{N}{An integer specifying the number of iterations for the empirical functions used.}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}
}
\value{
A series of highest density intervals for each parameter.
}
\description{
This function calculates the highest density interval for the parameters of a Gamma posterior, as with a Poisson response variable.
}
