% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_condition_setter.R
\name{initial_condition_setter}
\alias{initial_condition_setter}
\title{Initial Condition Setter}
\usage{
initial_condition_setter(
  data,
  n_leaves,
  p_0 = NULL,
  l_0 = NULL,
  initial_method = NULL,
  variable_time = TRUE
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if variable) are the last two columns.}

\item{n_leaves}{An integer specifying the number of distinct variable combinations in the data set.}

\item{p_0}{A numeric vector of initial values for the risk probabilities. If the vector is of length 1, this value will be repeated for each leaf. If NULL, then an automatic method (initial_method) will be used.}

\item{l_0}{A numeric vector of initial values for the rates. If the vector is of length 1, this value will be repeated for each leaf. If NULL, then an automatic method (initial_method) will be used.}

\item{initial_method}{A character string indicating the method for automatically setting the initial conditions. The character string can be an element of c("mean",mle","mm"), with "mean" being the default when p_0, l_0 are not provided.}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}
}
\value{
A list containing vectors of initial values for the risk probability and rat, as well as a logical value for whether mean is being used as the initial condition method (TRUE) or not (FALSE).
}
\description{
Sets the initial conditions for the nlm_zip() and em_zip() functions, based on inputs.
}
\examples{
initial_condition_setter(knee_pain_obs,8)

}
