% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{keeper}
\alias{keeper}
\title{The Keeper Function}
\usage{
keeper(v, n)
}
\arguments{
\item{v}{A numeric vector}

\item{n}{A numeric value}
}
\value{
A vector of the same length where values equal to n are 1, and values different to n are 0
}
\description{
A function that, when given a specific value, keeps all values in a vector that are equal as 1, and any value that is different as 0.
}
