% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{LabelStage1}
\alias{LabelStage1}
\title{The Label Stage Function}
\usage{
LabelStage1(k, num.variable, num.situation, label.category, num.category)
}
\arguments{
\item{k}{An integer value}

\item{num.variable}{An integer value}

\item{num.situation}{An integer value}

\item{label.category}{A string detailing the label of the category}

\item{num.category}{An integer value}
}
\value{
A vector of labels
}
\description{
The function \code{\link[ceg:LabelStage]{ceg::LabelStage()}}
}
