% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood_function_NOT_READY.R
\name{loglik}
\alias{loglik}
\title{Log Likelihood Calculator}
\usage{
loglik(mod)
}
\arguments{
\item{mod}{A StagedTree object.}
}
\value{
A list containing the log likelihood, number of stages and BIC
}
\description{
Calculates the log likelihood and BIC for a StagedTree object.
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
loglik(mod)
}
