% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effect_functions.R
\name{marginal_effect}
\alias{marginal_effect}
\title{The Marginal Effect Function}
\usage{
marginal_effect(
  data,
  mod,
  input_variable = c(),
  rel_output = 0,
  max_per_plot = 4,
  zip = FALSE
)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{mod}{A StagedTree model fit to the data set, as produced by pceg() or zipceg()}

\item{input_variable}{An integer vector detailing the covariates whose marginal effect is to be analysed. The default is to analyse all covariates.}

\item{rel_output}{A non-positive integer value detailing the variable which is being affected by the input variable(s), relative to the response. When 0, this will analyse the response variable, but any variable that appears after the input variables in the tree can be analysed.}

\item{max_per_plot}{An integer value specifying the maximum number of leaves that can be shown in a single plot.}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}
}
\value{
A marginal effect plot for each variable.
}
\description{
The Marginal Effect Function
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
marginal_effect(knee_pain_obs,mod)
}
