% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_divergence.R
\name{matrix_to_vector}
\alias{matrix_to_vector}
\title{The Matrix to Vector Function}
\usage{
matrix_to_vector(mat)
}
\arguments{
\item{mat}{A numeric matrix with m rows and n columns}
}
\value{
A numeric vector with mn elements, where each row of the matrix is concatenated to the next.
}
\description{
Convert a matrix into a vector, row by row.
}
\examples{
mat<-matrix(runif(6),nrow=3);mat
matrix_to_vector(mat)
}
