% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_probability_functions.R
\name{merge_prob}
\alias{merge_prob}
\title{The Merge Probability function}
\usage{
merge_prob(
  n,
  p1,
  p2 = NULL,
  l1,
  l2,
  alpha = NULL,
  a11 = NULL,
  a12 = NULL,
  a21 = NULL,
  a22 = NULL
)
}
\arguments{
\item{n}{An integer value for the total sample size.}

\item{p1}{A numeric value between 0 and 1 specifying the assumed true probability for the first stage.}

\item{p2}{A numeric value between 0 and 1 specifying the assumed true probability for the second stage. If NULL, p1 is used by default.}

\item{l1}{A numeric value between 0 and 1 specifying the path probability up to the first stage.}

\item{l2}{A numeric value between 0 and 1 specifying the path probability up to the second stage.}

\item{alpha}{A numeric value specifying the equivalent sample size used for setting the prior. If NULL, the inputted prior values are used.}

\item{a11}{A numeric value specifying the first prior component for the first stage. If NULL, alpha is used for the prior instead.}

\item{a12}{A numeric value specifying the second prior component for the first stage. If NULL, alpha is used for the prior instead.}

\item{a21}{A numeric value specifying the first prior component for the second stage. If NULL, alpha is used for the prior instead.}

\item{a22}{A numeric value specifying the second prior component for the second stage. If NULL, alpha is used for the prior instead.}
}
\value{
A list containing i) A matrix of the joint probabilities of each possible count combination; ii) A matrix of whether each count combination merges or not; iii) the total probability of merging, the sum of the elementwise product of (i) and (ii).
}
\description{
Calculate the probability of two true stages being merged using the AHC algorithm.
}
\details{
Given a sample size, at least one stage probability, path probabilities and a prior structure, returns the probability of the two stages being merged. If only one probability is specified, the second probability is assumed to be the same. If alpha is provided, the priors are based on a uniform path prior and the stage probabilities.
}
\examples{
merge_prob(100,p1=0.4,p2=0.6,l1=1/8,l2=1/8,alpha=1)
merge_prob(100,p1=0.4,p2=0.6,l1=1/8,l2=1/8,a11=1/16,a12=1/16,a21=1/16,a22=1/16)
merge_prob(100,p1=0.4,p2=0.6,l1=1/8,l2=1/8,a11=1/20,a12=3/40,a21=3/40,a22=1/20)
}
