% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_counts_function.R
\name{merge_separator}
\alias{merge_separator}
\title{The merge_Separator function}
\usage{
merge_separator(mod, n, p, tree, data_levels, zip = FALSE)
}
\arguments{
\item{mod}{An object of the S3 class StagedTree}

\item{n}{An integer value}

\item{p}{An integer value}

\item{tree}{A matrix}

\item{data_levels}{A numeric vector}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}
}
\value{
A list containing a matrix and two vectors.
}
\description{
The merge_Separator function
}
