% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_counts_function.R
\name{merged_list_extractor}
\alias{merged_list_extractor}
\title{The Merged List Extractor}
\usage{
merged_list_extractor(merged)
}
\arguments{
\item{merged}{A matrix specifying the merging present in a CEG.}
}
\value{
A list of intger vectors detailing the stages which have merged.
}
\description{
A function to find a list of the stages which have merged given the merging details of a CEG.
}
