% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_trees_functions.R
\name{model_combiner}
\alias{model_combiner}
\title{Model Combiner}
\usage{
model_combiner(data, mod.background, mod.response, prior.input = NULL)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if included) are the last two columns.}

\item{mod.background}{A StagedTree object modelling the background process.}

\item{mod.response}{A StagedTree object modelling the response process.}

\item{prior_input}{A list where each element is a matrix corresponding to a Dirichlet prior distribution for each level of the tree. If NULL, the default prior is used.}
}
\value{
A StagedTree object which is equivalent to combining mod.background and mod.response.
}
\description{
Takes a background model and response model and combines them in one full model.
}
\details{
Intended for use with Intermediate CEGs, this creates a StagedTree object which is equivalent to combining the background process in mod.background, with the response process in mod. response. This combined model can be used to compare to other methods of model selection, such as the score.
}
\examples{
mod1<-pceg(knee_pain_obs,2,TRUE,TRUE)
mod.back<-pceg(knee_pain_obs[,1:3],2,FALSE,FALSE)
newdata<-background_extractor(knee_pain_obs,mod.back)
mod.resp<-pceg(newdata,2,TRUE,TRUE)
mod2<-model_combiner(knee_pain_obs,mod.back,mod.resp)
}
