% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{nonzero_checker}
\alias{nonzero_checker}
\title{The Nonzero Checker}
\usage{
nonzero_checker(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
An integer value specifying how how many elements of the vector are zero.
}
\description{
A function that checks how many elements of a given vector are nonzero.
}
