% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pceg_function.R
\name{nonzero_vector_checker}
\alias{nonzero_vector_checker}
\title{Nonzero Vector Checker}
\usage{
nonzero_vector_checker(v)
}
\arguments{
\item{v}{A numerical vector}
}
\value{
An integer detailing how many elements are nonzero
}
\description{
Calculates how many elements in a vector are nonzero
}
\examples{
nonzero_vector_checker(c(1,0,2))
}
