% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_new_functions.R
\name{oahc_definer}
\alias{oahc_definer}
\title{The OAHC Definer}
\usage{
oahc_definer(cluster, score = 0)
}
\arguments{
\item{cluster}{A list detailing the stage structure for a level}

\item{score}{A numeric value of the score contribution from this level}
}
\value{
An object of class "OAHC" with a stage structure and score from a level of the tree
}
\description{
This function creates an object of class "OAHC" from a pceg model.
}
