% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_new_functions.R
\name{output_list_converter}
\alias{output_list_converter}
\title{The Output List Converter Function}
\usage{
output_list_converter(mod, poisson_response = TRUE)
}
\arguments{
\item{mod}{A list created in \code{\link[=pceg]{pceg()}}.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}
}
\value{
A list for the prior distribution and a list for the posterior distribution.
}
\description{
A function to transform the list of prior and data outputs into the S3 class versions of the prior and posterior.
}
