% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{PairwisePosition1}
\alias{PairwisePosition1}
\title{The Pairwise Position Function}
\usage{
PairwisePosition1(pair.situation, num.category, pos.next.level)
}
\arguments{
\item{pos.next.level}{}
}
\value{

}
\description{
The Pairwise Position Function
}
