% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_counts_function.R
\name{parameter_extractor}
\alias{parameter_extractor}
\title{The Parameter Extractor Function}
\usage{
parameter_extractor(
  stage_struct,
  posterior,
  var,
  poisson_response = TRUE,
  remove_risk_free = TRUE
)
}
\arguments{
\item{stage_struct}{A list detailing the stage structure from a StagedTree object.}

\item{posterior}{A list detailing the posterior expectations of parameters from a StagedTree object.}

\item{var}{An integer value indicating which variable should have its parameters extracted.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{remove_risk_free}{A logical value indicating whether the risk free leaves and edges should be removed (TRUE) or not (FALSE).}
}
\value{
A numeric matrix or vector (for Poisson rates) detailing the individual posterior expectations for each situation for the chosen variable.
}
\description{
The Parameter Extractor Function
}
