% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_divergence.R
\name{path_prob_extractor}
\alias{path_prob_extractor}
\title{The Path Probability Extractor}
\usage{
path_prob_extractor(mod, precision = 5)
}
\arguments{
\item{mod}{A StagedTree object.}

\item{precision}{An integer value specifying how many decimal places numerical checks should be done to.}
}
\value{
A list containing: a stages matrix, outlining the individual probability for each transition along a path, and a paths matrix, calculating the cumulative product of transition probabilities such that the final column is the probability of traversing each path.
}
\description{
Calculate the probability of traversing each path in a StagedTree object. For a StagedTree with a Poisson response, it calculates the probability of traversing the path until the leaf where the response is recorded.
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
path_prob_extractor(mod)
}
