% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_functons.R
\name{path_refactor}
\alias{path_refactor}
\title{The Path refactor function}
\usage{
path_refactor(data, n, data_levels)
}
\arguments{
\item{data}{A data set where the covariates are factors.}

\item{n}{An integer for the number of columns to refactor, beginning from the start.}

\item{data_levels}{An integer vector of the number of levels for each covariate.}
}
\value{
A data set where the categorical covariates have been refactor to start at 0 and end at the number of levels - 1.
}
\description{
This function takes the categorical covariates from a data set and refactors each one from 0 to the number of levels - 1.
}
\examples{
N<-100
h<-sample(c("Short","Average","Tall"),size=N,replace=TRUE)
h<-factor(h,levels=c("Short","Average","Tall"))
w<-sample(c("Normal","Over"),size=N,replace=TRUE)
w<-factor(w,levels=c("Normal","Over"))
y<-rpois(N,lambda=2)
data<-data.frame(height=h,weight=w,y=y)
summary(data)
n<-2
data_levels<-sapply(data[,1:n],nlevels)
data.refactor<-path_refactor(data,n=2,data_levels)
summary(data.refactor)
}
