% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pceg_function.R
\name{pceg}
\alias{pceg}
\title{The PCEG function}
\usage{
pceg(
  data,
  equivsize = 3,
  poisson_response = FALSE,
  variable_time = FALSE,
  zip = FALSE,
  gamma_alpha = 1,
  gamma_beta = 2,
  structural_zero = FALSE,
  var_disc = 0,
  disc_length = 0,
  restrict = FALSE,
  mirror = FALSE,
  cat_limit = 0
)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{equivsize}{A numeric value specifying the equivalent sample size for the prior, a measure of confidence in the prior.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}

\item{gamma_alpha}{A numeric value for the shape hyperparameter of the Gamma prior for the Poisson rate, if applicable.}

\item{gamma_beta}{A numeric value for the rate hyperparameter of the Gamma prior for the Poisson rate.}

\item{structural_zero}{A logical value indicating whether zero counts in the data set should be considered as structural (TRUE) or sampling (FALSE) for the setting of the prior.}

\item{var_disc}{An integer value specifying which variable to discretise. If 0, no discretisation is necessary.}

\item{disc_length}{An integer value specifying how many neighbours can be searched over for the purposes of variable discetisation. If 0, all other possible stages may be merged over.}

\item{restrict}{A logical value indicating whether variable discretisation should be restricted to stages with the same unfolding of the process (TRUE) or not (FALSE).}

\item{mirror}{A logical value indicating whether variable discretisation should be equivalent across each unfolding of the process (TRUE) or not (FALSE).}

\item{cat_limit}{An integer value specifying the minimum number of categories to the variable can be discretised to. If 0, there is no minimum number of categories.}
}
\value{
A list specifying a PCEG model. The list contains: the prior for the final model, the data for the final model, the stage numbers for the final model,
the stage structure for the final model, the vector of likelihoods after each merging, the details of the stages merged at each step,
the comparison set of stages left to be merged, and the log marginal likelihood of the final model.
}
\description{
This function fits a PCEG model to a data set.
}
\details{
This function takes a data set, a prior structure, and various other inputs to fit a PCEG model. It is flexible enough to fit a vanilla CEG, and perform variable discretisation methods.
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
mod$result
}
