% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions_S3.R
\name{plot.ChainEventGraph}
\alias{plot.ChainEventGraph}
\title{The \code{\link[=plot]{plot()}} method for the S3 Class ChainEventGraph}
\usage{
plot.ChainEventGraph(x)
}
\arguments{
\item{x}{An object in the S3 class ChainEventGraph}
}
\value{
A plot of the ChainEventGraph
}
\description{
The \code{\link[=plot]{plot()}} method for the S3 Class ChainEventGraph
}
