% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions_S3.R
\name{plot.EventTree}
\alias{plot.EventTree}
\title{The \code{\link[=plot]{plot()}} method for the S3 Class EventTree}
\usage{
plot.EventTree(x, ...)
}
\arguments{
\item{x}{An object in the S3 class EventTree}
}
\value{
A plot of the EventTree
}
\description{
The \code{\link[=plot]{plot()}} method for the S3 Class EventTree
}
