% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pceg_function.R
\name{poisson_score}
\alias{poisson_score}
\title{The Poisson Score}
\usage{
poisson_score(data_sample, gamma_prior)
}
\arguments{
\item{data_sample}{A list where each element of the list contains the observed count vector and time vector for a unique unfolding of the process.}

\item{gamma_prior}{A list of numeric vectors where the numeric vectors are the prior hyperparameters for the Gamma distribution.}
}
\value{
A numeric value of the log marginal likelihood for the data.
}
\description{
This function calculates the marginal likelihood for Poisson responses.
}
