% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{PositionLevel1}
\alias{PositionLevel1}
\title{The Position Level Function}
\usage{
PositionLevel1(
  stage.list,
  num.category,
  num.situation.next,
  pos.next.level = list()
)
}
\arguments{
\item{pos.next.level}{}
}
\value{

}
\description{
The Position Level Function
}
