% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{PositionStage1}
\alias{PositionStage1}
\title{The Position Stage Function}
\usage{
PositionStage1(stage.vector, num.category, pos.next.level)
}
\arguments{
\item{stage.vector}{An integer vector detailing the stages merged at a single level.}

\item{num.category}{An integer value for the number of categories for the associated variable.}

\item{pos.next.level}{An integer vector detailing the position structure for a single level.}
}
\value{
A list
}
\description{
The function \code{\link[ceg:PositionStage]{ceg::PositionStage()}}
}
