% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{PositionVector1}
\alias{PositionVector1}
\title{The Position Vector Function}
\usage{
PositionVector1(num.situation, pos.list)
}
\arguments{
\item{num.situation}{An integer value detailing the number of situations at a single level.}

\item{pos.list}{A list detailing the position structure for a single level.}
}
\value{
An integer vector detailing the position structure for a single level.
}
\description{
The function \code{\link[ceg:PositionVector]{ceg::PositionVector()}}
}
