% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_function.R
\name{prior_set}
\alias{prior_set}
\title{The Prior Set Function}
\usage{
prior_set(sample, struc_zero, numbvariables, numbcat, numb, equivsize)
}
\arguments{
\item{sample}{A list containing the edge counts of the situations in a data set.}

\item{struc_zero}{A logical value indicating whether zero edge counts in the sample should be considered as structural zeroes (TRUE) or sampling zeroes (FALSE) for the purposes of setting the prior.}

\item{numbvariables}{An integer value specifying how many covariates are in the data set, or levels in the event tree.}

\item{numbcat}{An integer vector specifying the number of categories for each covariate in the data set, or outgoing edges at each level of the event tree.}

\item{numb}{An integer vector specifying the number of situations at each level of the event tree.}

\item{equivsize}{A numeric value specifying the equivalent sample size for the prior, a measure of confidence in the prior.}
}
\value{
A list containing the prior hyperparameters for each situation and its outgoing edges in the event tree.
}
\description{
This function sets the prior for the situations in the tree.
}
\details{
This function sets the prior for the situations in the tree which have categorical response variables, based on the equivalent sample size and how zero edge counts should be treated.
}
