% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_tree_functions.R
\name{quantile_band}
\alias{quantile_band}
\title{The Quantile Band Function}
\usage{
quantile_band(
  data,
  mod,
  signif = 0.05,
  limit = NA,
  shift = TRUE,
  poisson_response = TRUE,
  variable_time = TRUE,
  zip = FALSE
)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{mod}{A StagedTree model fit to the data set, as produced by pceg() or zipceg()}

\item{signif}{A numeric value specifying the significance level for the quantiles.}

\item{limit}{An integer vector specifying the maximium number of possible counts to analyse. If it is a vector of length one, this value will be used for all leaves. If NA, this will be the maximum count recorded per leaf.}

\item{shift}{A logical value indicating whether the raw observed counts and quantiles should be used (FALSE), or whether they should be shifted by the median (TRUE).}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}
}
\value{
A quantile-band plot for each leaf.
}
\description{
This function creates a quantile band plot, a diagnostic plot of the suitability of a model to count data.
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
quantile_band(knee_pain_obs,mod,limit=10,zip=FALSE)
}
