% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_matrix_function.R
\name{refactored_tree_matrix}
\alias{refactored_tree_matrix}
\title{The Tree Matrix Function}
\usage{
refactored_tree_matrix(data, variable_time)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}
}
\value{
A list containing the refactored data, the number of levels in the tree, the number of leaves in the tree, a vector of the levels for each variable, and the refactored tree matrix.
}
\description{
A function to refactor the data and create a refactored matrix that matches the covariates in the tree.
}
\examples{
refactored_tree_matrix(knee_pain_obs,TRUE)
}
