% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_trees_functions.R
\name{saturated_model_computer}
\alias{saturated_model_computer}
\title{Saturated Model Computer}
\usage{
saturated_model_computer(
  data,
  cutpoint.variable = NULL,
  equivsize = 2,
  poisson_response = FALSE,
  variable_time = FALSE,
  gamma_alpha = 1,
  gamma_beta = 2,
  prior_input = NULL
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if included) are the last two columns.}

\item{cutpoint.variable}{An integer value specifying after which variable the tree should be saturated. If NULL, the entire tree will be saturated.}

\item{equivsize}{A numeric value specifying the equivalent sample size for the prior, a measure of confidence in the prior.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{gamma_alpha}{A numeric vector for the shape hyperparameters of the Gamma priors for the Poisson rates, if applicable. If a single value, each prior takes the same value.}

\item{gamma_beta}{A numeric vector for the rate hyperparameters of the Gamma priors for the Poisson rates, if applicable. If a single value, each prior takes the same value.}

\item{prior_input}{A list where each element is a matrix corresponding to a Dirichlet prior distribution for each level of the tree.}
}
\value{
A StagedTree model where the variables up to the cutpoint variable are merged normally, and any subsequent variables are saturated.
}
\description{
A function to quickly compute a saturated StagedTree model from a data set, with a cutpoint variable specifying where the saturation begins.
}
\examples{
saturated_model_computer(knee_pain_obs,3,2,TRUE,TRUE)
}
