% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{sceg}
\alias{sceg}
\title{The Stratified CEG Function}
\usage{
sceg(stratified.staged.tree)
}
\arguments{
\item{stratified.staged.tree}{An object of class "Stratified.staged.tree"}
}
\value{
An object of class "Stratified.ceg.model"
}
\description{
This is the \code{\link[ceg:Stratified.event.tree]{ceg::Stratified.event.tree()}} constructor method.
}
\examples{
mod1<-pceg(knee_pain_obs,2,TRUE,TRUE)
 staged.tree<-staged.tree.creator(mod1)
 mod2<-sceg(staged.tree)
 plot(mod2)
}
