% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{set}
\alias{set}
\title{The Stratified Event Tree Function}
\usage{
set(x = "data.frame", ...)
}
\arguments{
\item{x}{A data frame}

\item{...}{}
}
\value{
A Statified.event.tree object
}
\description{
This is a slightly modified version of the \code{\link[ceg:stratified.event.tree]{ceg::stratified.event.tree()}} constructor method.
}
\examples{
set(knee_pain_obs[,1:3])
}
