% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/situation_stage_monitor.R
\name{situation_stage_monitor}
\alias{situation_stage_monitor}
\title{The Situation Stage monitor}
\usage{
situation_stage_monitor(
  data,
  mod,
  var = NULL,
  stage = NULL,
  cat = 1,
  N = 10000,
  ci = 0.95,
  zip = FALSE
)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{mod}{A StagedTree model fit to the data set, as produced by \code{\link[=pceg]{pceg()}}.}

\item{var}{An integer value detailing which variable, if applicable.}

\item{stage}{An integer value detailing either which stage in the tree to investigate, or which stage for a chosen variable.}

\item{cat}{An integer value detailing which category of the applicable variable should be used as reference.}

\item{N}{An integer specifying the number of iterations for the empirical functions used.}

\item{ci}{A numeric value between 0 and 1 specifying the percentage confidence for the highest density interval.}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}
}
\value{
A plot of confidence intervals versus observed when each situation in the stage is removed.
}
\description{
A function to check if any situations in a stage are different to the rest.
}
\examples{
mod<-pceg(knee_pain_obs,2)
situation_stage_monitor(knee_pain_obs,mod,var=4,stage=3) #the third stage for variable 4
situation_stage_monitor(knee_pain_obs,mod,stage=8) #the 10th stage in the tree, which is the same as the above
}
