% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_trees_functions.R
\name{stage_updater}
\alias{stage_updater}
\title{Stage Updater}
\usage{
stage_updater(mod, level, ref1, ref2)
}
\arguments{
\item{mod}{A StagedTree object.}

\item{level}{An integer value specifying what level of the tree the merging takes place at.}

\item{ref1}{An integer value specifying what the first stage at the chosen level to be merged is.}

\item{ref2}{An integer value specifying what the second stage at the chosen level to be merged is.}
}
\value{
A StagedTree object which has merged the two stages.
}
\description{
Updates a model by merging two stages.
}
\examples{
mod<-pceg(knee_pain_obs,2);plot(mod)
mod2<-stage_updater(mod,2,1,2);plot(mod2)
mod3<-stage_updater(mod,3,1,3);plot(mod3)
}
