% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_new_functions.R
\name{staged.tree.creator}
\alias{staged.tree.creator}
\title{The Staged Tree Creator}
\usage{
staged.tree.creator(
  data,
  mod,
  poisson_response = TRUE,
  variable_time = TRUE,
  zip = FALSE,
  score = 0
)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}

\item{score}{A numeric vector of the score contributions from each level}
}
\value{
An object of the class "Stratified.staged.tree"
}
\description{
#' A function to create an object of the "Stratified.event.tree" S4 class that is compatible with the outputs of the \code{\link[=pceg]{pceg()}} function.
}
\examples{
 mod1<-pceg(knee_pain_obs,2,TRUE,TRUE)
 staged.tree<-staged.tree.creator(mod1)
 plot(staged.tree)
}
