% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions_S3.R
\name{StagedTree}
\alias{StagedTree}
\title{The StagedTree constructor function}
\usage{
StagedTree(
  data,
  prior,
  counts,
  posterior,
  stage.struc,
  stages,
  merged,
  result,
  poisson_response = TRUE,
  variable_time = TRUE,
  zip = FALSE,
  remove_risk_free = FALSE,
  lik = 0
)
}
\arguments{
\item{data}{A data set where the observed response vector and time vector (if applicable and variable) are the last two columns.}

\item{prior}{A list detailing the prior distribution for the model.}

\item{counts}{A list specifying the edge counts or event counts and times for the model.}

\item{posterior}{A list specifying the posterior expectations for the model.}

\item{stage.struc}{A list specifying the stage structure for the model.}

\item{stages}{A numeric vector indicating which of the original situations are now stages.}

\item{merged}{A matrix specifying which situations have been merged into stages.}

\item{result}{A list detailing the covariate values in each stage.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{zip}{A logical value indicating whether the process is zero-inflated (TRUE) or not (FALSE). If the process is zero-inflated but this has already been accounted for in the data input, this should still be FALSE.}

\item{remove_risk_free}{A logical value indicating whether the risk free leaves and edges should be removed (TRUE) or not (FALSE).}

\item{lik}{A numeric value specifying the log marginal likelihood for the model.}
}
\value{
An object of the S3 class StagedTree
}
\description{
Create an object in the StagedTree S3 class from certain inputs for a model. Some fo these inputs can be calculated directly in the \code{\link[=pceg]{pceg()}} function.
}
