% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_functions.R
\name{state_imputer}
\alias{state_imputer}
\title{The State Imputer}
\usage{
state_imputer(
  data,
  lambda = 1,
  prob = 0.5,
  variable_time = TRUE,
  stoch = TRUE,
  all_risk = FALSE
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if variable) are the last two columns}

\item{lambda}{A numeric vector of the Poisson rates used to impute the states. If the vector is of length 1, this value will be repeated for each leaf.}

\item{prob}{A numeric vector of the risk probabilities used to impute the states.If the vector is of length 1, this value will be repeated for each leaf.}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}

\item{stoch}{A logical value indicating whether stochastic (TRUE) or deterministic (FALSE) imputation should be used.}

\item{all_risk}{A logical value indicating whether all observations should be considered at risk. This is necessary when converting a PCEG to a ZIPCEG for the purposes of calculating the Bayes Factor.}
}
\value{
A data set containing refactored covariates, imputed risk states, and observed counts and times.
}
\description{
This function imputes the latent risk states based on the data, the estimated parameters, and the choice of imputation method.
}
\examples{
state_imputer(knee_pain_obs)
}
