% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{StratifiedCegPosition1}
\alias{StratifiedCegPosition1}
\title{The StratifiedCEGPosition1 Function}
\usage{
StratifiedCegPosition1(stage, num.category, num.situation)
}
\arguments{
\item{stage}{A list detailing the stage structure of a Stratified.staged.tree model at each level.}

\item{num.category}{An integer vector detailing the number of categories for each variable.}

\item{num.situation}{An integer vector detailing the number of situations at each level.}
}
\value{
A list detailing the position structure at each level.
}
\description{
The function \code{\link[ceg:StratifiedCEGPosition]{ceg::StratifiedCEGPosition()}}
}
