% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions_S3.R
\name{summary.ChainEventGraph}
\alias{summary.ChainEventGraph}
\title{The \code{\link[=summary]{summary()}} method for the S3 Class ChainEventGraph}
\usage{
\method{summary}{ChainEventGraph}(object)
}
\arguments{
\item{object}{An object in the S3 class ChainEventGraph}
}
\value{
A summary of the StagedTree object within the ChainEventGraph object, displaying the makeup of each stage, the posterior expectations, and the score.
}
\description{
The \code{\link[=summary]{summary()}} method for the S3 Class ChainEventGraph
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
ceg<-ChainEventGraph(mod)
summary(ceg)
}
