% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions_S3.R
\name{summary.StagedTree}
\alias{summary.StagedTree}
\title{The \code{\link[=summary]{summary()}} method for the S3 Class StagedTree}
\usage{
\method{summary}{StagedTree}(object)
}
\arguments{
\item{object}{An object in the S3 class StagedTree}
}
\value{
A summary of the StagedTree object, displaying the makeup of each stage, the posterior expectations, and the score.
}
\description{
The \code{\link[=summary]{summary()}} method for the S3 Class StagedTree
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
summary(mod)
}
