% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{time_minimiser}
\alias{time_minimiser}
\title{The time minimiser}
\usage{
time_minimiser(data, t_min)
}
\arguments{
\item{data}{A data set where the last column is the observed time vector.}

\item{t_min}{A numeric minimum time value.}
}
\value{
A data set where the observed times now have a minimum value.
}
\description{
A function that enforces a minimum value of time for an observed time.
}
\examples{
n<-10
t<-runif(n)
y<-rpois(n,2*t)
data<-data.frame(y=y,t=t)
data
data.min<-time_minimiser(data,t_min=0.5)
data.min
}
