% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_extractor_function.R
\name{total_value_extractor}
\alias{total_value_extractor}
\title{The Total Expected Value Extractor}
\usage{
total_value_extractor(
  data,
  ceg,
  level_exclude = NA,
  poisson_response = TRUE,
  variable_time = TRUE,
  posterior = TRUE,
  zip = TRUE,
  dec_place = NA
)
}
\arguments{
\item{data}{A data set, where the observed response vector and time vector (if applicable and variable) are the last two columns}

\item{ceg}{A ceg model fit to the data set, as produced by pceg().}

\item{level_exclude}{A non-positive integer integer indicating which variables (if any) should be excluded from the output, is relative to the response variable.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{posterior}{A logical value indicating whether the estimates of the posterior (TRUE) or sample (FALSE) expected value should be calculated.}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}

\item{dec_place}{An integer value detailing how many decimal places the outputs should be rounded to. If NA, no rounding will occur.}
}
\value{

}
\description{
This function prints the expected values of parameters based on the data set and a chosen PCEG model for each variable.
}
\details{
This function is the \code{\link[=value_extractor]{value_extractor()}} function repeated at each level.
}
