% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plotter.R
\name{trace_plotter}
\alias{trace_plotter}
\title{The Traceplotter Function}
\usage{
trace_plotter(y_chain, y_est, ind = c(), max_per_plot = 8)
}
\arguments{
\item{y_chain}{A list where each element is a vector corresponding to the chain used to estimate a parameter.}

\item{y_est}{A vector with the final estimate for each parameter. This will appear as a red horizontal line.}

\item{ind}{A vector of which chains to plot. If empty, this will produce a trace plot of each chain.}

\item{max_per_plot}{An integer value specifying the maximum number of chains that can be shown in a single plot.}
}
\value{
A traceplot for each chain provided, along with the final estimate as a red horizontal line.
}
\description{
Produces a traceplot for chains produced in order to estimate the underlying parameters of the ZIP distribution.
}
\examples{
gibbs_result<-gibbs_zip(knee_pain_obs)
l<-gibbs_result$lambda
l_est<-gibbs_result$summary$l_hat
trace_plotter(l,l_est)
trace_plotter(l,l_est,ind=c(1,5))
}
