% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_divergence.R
\name{tree_repeater}
\alias{tree_repeater}
\title{The Tree Repeater Function}
\usage{
tree_repeater(vec, num_leaves)
}
\arguments{
\item{vec}{A numeric vector of stage probabilities.}

\item{num_leaves}{An integer count of the number of leaves in the StagedTree object.}
}
\value{
A numeric vector the length of num_leaves where each element is the corresponding stage probability for that part of the path.
}
\description{
Repeat stage probabilities for each path they are a constituent of.
}
\examples{
tree_repeater(c(0.5,0.5),8)
}
