% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{TruncatedPath1}
\alias{TruncatedPath1}
\title{The Label Stage Function}
\usage{
TruncatedPath1(ref, k, var, num.category, num.situation, label.category)
}
\arguments{
\item{label.category}{}
}
\value{

}
\description{
The function \link[ceg:TruncatedPath]{ceg::TruncatedPath}
}
