% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_time_functions.R
\name{uniform_time_zip}
\alias{uniform_time_zip}
\title{The Uniform Time ZIP Estimator}
\usage{
uniform_time_zip(data, method = "mle", time_input = FALSE)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if included) are the last two columns.}

\item{method}{A character string indicating the method for parameter estimation. The character string can be an element of c("mle","mm").}

\item{time_input}{A logical value indicating whether the time has been included in the data set (TRUE) or not (FALSE). As time is uniform, if it is included, then all times should be equal to 1.}
}
\value{
A list containing a matrix with the desired outputs for each evolution of the process, as well as vectors of the estimates for the rates and risk probabilities.
}
\description{
This function can estimate the risk proportions and rates for counts with uniform time, using either a maximum likelihood or methods of moments estimator.
}
\examples{
nlm_zip(knee_pain_obs[,1:4])
}
