% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{vec_from}
\alias{vec_from}
\title{The vec_from function}
\usage{
vec_from(x)
}
\arguments{
\item{x}{A positive integer}
}
\value{
A vector of integers from $x-1$ to 0.
}
\description{
A function to create a descending vector of natural numbers
}
\examples{
vec_from(7)
}
