% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_counts_function.R
\name{vec_in}
\alias{vec_in}
\title{The vec_in function}
\usage{
vec_in(x, v)
}
\arguments{
\item{x}{A numeric value}

\item{v}{A numeric vector}
}
\value{
A logical value indicating whether the given value was in the vector (TRUE) or not (FALSE)
}
\description{
This function checks whether a given value is in a vector.
}
