% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{vec_to}
\alias{vec_to}
\title{The vec_to function}
\usage{
vec_to(x)
}
\arguments{
\item{x}{A positive integer}
}
\value{
A vector of integers from 0 to $x-1$.
}
\description{
A function to create an ascending vector of natural numbers
}
