% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\title{The variance inflation factor calculation for a level of a staged tree}
\usage{
vif(mod, var)
}
\arguments{
\item{mod}{A StagedTree model fit to the data set, as produced by pceg() or zipceg().}

\item{var}{A positive integer value detailing which level of the tree to calculate the VIF for.}
}
\value{
A numeric value for the variance inflation factor.
}
\description{
The variance inflation factor calculation for a level of a staged tree
}
\examples{
mod1<-pceg(knee_pain_obs,2,TRUE,TRUE)
vif(mod1,3)
}
