% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{zero_checker}
\alias{zero_checker}
\title{The Zero Checker}
\usage{
zero_checker(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
An integer value specifying how how many elements of the vector are zero.
}
\description{
A function that checks how many elements of a given vector are zero.
}
