% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_functions.R
\name{zipceg.iter}
\alias{zipceg.iter}
\title{The Iterative ZIPCEG function}
\usage{
zipceg.iter(
  data,
  method = "Gibbs",
  iter_total = 10,
  iter_f = 10000,
  plot_rates = TRUE,
  plot_probs = FALSE,
  hist = FALSE,
  line = FALSE,
  scatter = FALSE,
  equivsize = 2,
  poisson_response = TRUE,
  variable_time = FALSE,
  stoch_imputation = TRUE,
  print_output = FALSE,
  gamma_alpha = 1,
  gamma_beta = 2,
  beta_c = 1,
  beta_d = 1,
  p_0 = NA,
  l_0 = NA,
  tol = 1e-10,
  var_disc = 0,
  disc_length = 0,
  restrict = FALSE,
  mirror = FALSE,
  cat_limit = 0
)
}
\arguments{
\item{data}{A data set where the observed count vector and time vector (if variable) are the last two columns.}

\item{method}{A character string indicating the method for parameter estimation. The character string can be an element of c("Gibbs","nlm","EM","mle","mm").}

\item{iter_total}{An integer specifying the number of iterations the model selection process should be performed for.}

\item{iter_f}{An integer specifying the number of iterations for the parameter estimation method chosen, if necessary.}

\item{plot_rates}{A logical value indicating whether the stage rates should be plotted (TRUE) or not (FALSE).}

\item{plot_probs}{A logical value indicating whether the risk probabilities should be plotted (TRUE) or not (FALSE).}

\item{hist}{A logical value indicating whether the plot should be in the form of a histogram (TRUE) or not (FALSE).}

\item{line}{A logical value indicating whether the plot should be in the form of a line plot (TRUE) or not (FALSE).}

\item{scatter}{A logical value indicating whether the plot should be in the form of a scatter plot (TRUE) or not (FALSE).}

\item{equivsize}{A numeric value specifying the equivalent sample size for the prior, a measure of confidence in the prior.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{stoch_imputation}{A logical value indicating whether stochastic (TRUE) or deterministic (FALSE) imputation should be used.}

\item{print_output}{A logical value indicating whether the output for each iteration should be printed (TRUE) or not (FALSE).}

\item{gamma_alpha}{A numeric value for the shape hyperparameter of the Gamma prior for the Poisson rate, if applicable.}

\item{gamma_beta}{A numeric value for the rate hyperparameter of the Gamma prior for the Poisson rate.}

\item{beta_c}{A numeric value for the alpha hyperparameter of the Beta prior for the risk probability.}

\item{beta_d}{A numeric value for the beta hyperparameter of the Beta prior for the risk probability.}

\item{p_0}{A numeric initial value for the risk probability.}

\item{l_0}{A numeric initial value for the rate.}

\item{tol}{A numeric which represents the minimum change in the complete data log likelihood needed to continue the Expectation-Maximisation algorithm.}

\item{var_disc}{An integer value specifying which variable to discretise. If 0, no discretisation is necessary.}

\item{disc_length}{An integer value specifying how many neighbours can be searched over for the purposes of variable discetisation. If 0, all other possible stages may be merged over.}

\item{restrict}{A logical value indicating whether variable discretisation should be restricted to stages with the same unfolding of the process (TRUE) or not (FALSE).}

\item{mirror}{A logical value indicating whether variable discretisation should be equivalent across each unfolding of the process (TRUE) or not (FALSE).}

\item{cat_limit}{An integer value specifying the minimum number of categories to the variable can be discretised to. If 0, there is no minimum number of categories.}
}
\value{
A list containing: a matrix of the estimated rates or risk probabilities for each leaf across iterations, a numeric value of the log marginal likelihood for the MAP model, and the MAP StagedTree model itself.
}
\description{
This function performs the ZIPCEG model selection a specified number of times and can produce plots and perform model selection.
}
\details{
This function takes the same inputs as the \code{\link[=zipceg]{zipceg()}} function, except carries out the model selection process a specified number of times. Using the results of each iteration, different types of plots can be used to display either the estimated risk probabilities or estimated rates. The default plot is a violin plot. The function will also select the \emph{maximum} \emph{a} \emph{posteriori} (MAP) model from these iterations.
}
\examples{
zipceg.iter(knee_pain_obs,"nlm",iter_total=100,variable_time=TRUE)
}
