% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_functions.R
\name{zipceg}
\alias{zipceg}
\title{The ZIPCEG function}
\usage{
zipceg(
  data,
  method = "Gibbs",
  iter = 10000,
  equivsize = 2,
  poisson_response = TRUE,
  variable_time = TRUE,
  remove_risk_free = TRUE,
  stoch_imputation = TRUE,
  gamma_alpha = 1,
  gamma_beta = 2,
  beta_c = 1,
  beta_d = 1,
  p_0 = NULL,
  l_0 = NULL,
  initial_condition = NULL,
  tol = 1e-10,
  structural_zero = FALSE,
  indep = NA,
  saturated = NA,
  mean_post_cluster = FALSE,
  var_disc = 0,
  disc_length = 0,
  restrict = FALSE,
  mirror = FALSE,
  cat_limit = 0
)
}
\arguments{
\item{data}{A data set where the observed count vector and time vector (if variable) are the last two columns.}

\item{method}{A character string indicating the method for parameter estimation. The character string can be an element of c("Gibbs","nlm","EM","mle","mm").}

\item{iter}{The number of iterations for the Gibbs sampler or Expectation-Maximisation algorithm.}

\item{equivsize}{A numeric value specifying the equivalent sample size for the prior, a measure of confidence in the prior.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{remove_risk_free}{A logical value indicating whether the risk free leaves and edges should be removed (TRUE) or not (FALSE).}

\item{stoch_imputation}{A logical value indicating whether stochastic (TRUE) or deterministic (FALSE) imputation should be used.}

\item{gamma_alpha}{A numeric value for the shape hyperparameter of the Gamma prior for the Poisson rate, if applicable.}

\item{gamma_beta}{A numeric value for the rate hyperparameter of the Gamma prior for the Poisson rate.}

\item{beta_c}{A numeric value for the alpha hyperparameter of the Beta prior for the risk probability.}

\item{beta_d}{A numeric value for the beta hyperparameter of the Beta prior for the risk probability.}

\item{p_0}{A numeric vector of initial values for the risk probabilities, if em or nlm method is chosen. If the vector is of length 1, this value will be repeated for each leaf. If NULL, then an automatic method (initial_condition) will be used.}

\item{l_0}{A numeric vector of initial values for the rates, if em or nlm method is chosen. If the vector is of length 1, this value will be repeated for each leaf. If NULL, then an automatic method (initial_condition) will be used.}

\item{initial_condition}{A character string indicating the method for automatically setting the initial conditions for the em and nlm methods. The character string can be an element of c("mean",mle","mm"), with "mean" being the default when p_0, l_0 are not provided.}

\item{tol}{A numeric which represents the minimum change in the complete data log likelihood needed to continue the Expectation-Maximisation algorithm.}

\item{structural_zero}{A logical value indicating whether zero counts in the data set should be considered as structural (TRUE) or sampling (FALSE) for the setting of the prior.}

\item{indep}{An integer vector indicating which variables should be assumed to be independent of preceding variables (all situations in same stage).}

\item{saturated}{An integer vector indicating which variables should be assumed to be saturated (all situations in own stage).}

\item{mean_post_cluster}{A logical value indicating whether mean posterior clustering should be used (TRUE) or not (FALSE).}

\item{var_disc}{An integer value specifying which variable to discretise. If 0, no discretisation is necessary.}

\item{disc_length}{An integer value specifying how many neighbours can be searched over for the purposes of variable discetisation. If 0, all other possible stages may be merged over.}

\item{restrict}{A logical value indicating whether variable discretisation should be restricted to stages with the same unfolding of the process (TRUE) or not (FALSE).}

\item{mirror}{A logical value indicating whether variable discretisation should be equivalent across each unfolding of the process (TRUE) or not (FALSE).}

\item{cat_limit}{An integer value specifying the minimum number of categories to the variable can be discretised to. If 0, there is no minimum number of categories.}
}
\value{
An object of the S3 class StagedTree.
}
\description{
This function fits a ZIPCEG model to the chosen data set, based on chosen methods of parameter estimation and state imputation.
}
\details{
This function takes the same inputs as the \code{\link[=pceg]{pceg()}}, along with chosen methods of parameter estimation and state imputation, to fit a ZIPCEG model. As with \code{\link[=pceg]{pceg()}} it can perform variable discretisation methods, but is not to be used for vanilla CEGs.
}
\examples{
mod<-zipceg(knee_pain_obs,"nlm")
plot(mod)
summary(mod)
}
