% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_optimisation_functions.R
\name{ziplike}
\alias{ziplike}
\title{The zero-inflated Poisson likelihood}
\usage{
ziplike(params, y, t)
}
\arguments{
\item{params}{A numeric vector of the parameters for the distribution.}

\item{y}{An integer vector of observed event counts.}

\item{t}{A numeric vector of observed times.}
}
\value{
A numeric of the zero-inflated Poisson log likelihood.
}
\description{
This function calculates the zero-inflated Poisson log likelihood for a count vector and time vector, given values of the parameters of a zero-inflated Poisson distribution.
}
