% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm_functions.R
\name{cdll}
\alias{cdll}
\title{The complete data log likelihood for a Poisson response}
\usage{
cdll(params, y, t)
}
\arguments{
\item{params}{A numeric vector of the parameters for the distribution.}

\item{y}{An integer vector of observed event counts.}

\item{t}{A numeric vector of observed times.}
}
\value{
A list containing the likelihood value and a vector of weights for the em algorithm
}
\description{
This function calculates the complete data log likelihood for a count vector and time vector, given values of the parameters of a zero-inflated Poisson distribution.
}
