% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm_functions.R
\name{em}
\alias{em}
\title{The EM algorithm for a leaf}
\usage{
em(p_0, l_0, y, t, tol = 1e-10, max_iter = 10000)
}
\arguments{
\item{p_0}{The initial value for the risk probability}

\item{l_0}{The initial value for the rate}

\item{y}{The observed count vector}

\item{t}{The observed time vector}

\item{tol}{The tolerance to stop the algorithm}

\item{max_iter}{The maximum number of iterations for the algorithm}
}
\value{
A list containing the estimates of the risk probability and the rate, and the number of iterations necessary
}
\description{
This function carries out the Expectation-Maximisation algorithm for a leaf, given initial values of the parameters and the count and time vectors.
}
