% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm_functions.R
\name{em_zip}
\alias{em_zip}
\title{The EM algorithm}
\usage{
em_zip(
  data,
  p_0 = 0.5,
  l_0 = 1,
  variable_time = TRUE,
  max_iter = 10000,
  tol = 1e-10
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if variable) are the last two columns.}

\item{p_0}{A numeric vector of initial values for the risk probabilities. If the vector is of length 1, this value will be repeated for each leaf.}

\item{l_0}{A numeric vector of initial values for the rates. If the vector is of length 1, this value will be repeated for each leaf}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}

\item{max_iter}{An integer for the maximum number of iterations for the algorithm.}

\item{tol}{A numeric which represents the minimum change in the complete data log likelihood needed to continue the algorithm.}
}
\value{
A list containing a matrix with the desired outputs for each evolution of the process, as well as vectors of the estimates for the rates and risk probabilities.
}
\description{
This function carries out the Expectation-Maximisation algorithm across all leaves.
}
\details{
For each leaf, which represents a unique evolution of the process, the \code{\link[=em]{em()}} function is applied.
}
\examples{
em_zip(knee_pain_obs)
}
