% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm_functions.R
\name{em_zip}
\alias{em_zip}
\title{The EM algorithm}
\usage{
em_zip(
  data,
  p_0 = NA,
  l_0 = NA,
  variable_time = TRUE,
  max_iter = 10000,
  tol = 1e-10
)
}
\arguments{
\item{data}{The entire data set, where the observed count vector and time vector are the last two columns}

\item{p_0}{The initial value for the risk value}

\item{l_0}{The initial value for the rate. If NA}

\item{variable_time}{An indicator of whether the observed time is uniform or variable}

\item{max_iter}{The tolerance to stop the algorithm}

\item{tol}{The maximum number of iterations for the algorithm}
}
\value{

}
\description{
This function carries out the Expectation-Maximisation algorithm across all leaves.
}
