% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_optimisation_functions.R
\name{f}
\alias{f}
\title{The zero-inflated poisson likelihood for a single observation.}
\usage{
f(p, lambda, y, t)
}
\arguments{
\item{p}{A numeric value between 0 and 1}

\item{lambda}{A numeric value}

\item{y}{An integer}

\item{t}{A positive numeric value}
}
\value{
A numeric value
}
\description{
The zero-inflated poisson likelihood for a single observation.
}
\examples{
f(0.5,1,1,1)
}
