% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_counts_function.R
\name{merged_stage_finder}
\alias{merged_stage_finder}
\title{The Merged Stage Finder}
\usage{
merged_stage_finder(unique_stage, M)
}
\arguments{
\item{unique_stage}{An integer value}

\item{M}{An integer matrix detailing the merging present in a CEG.}
}
\value{
An integer vector detailing the stages that have merged with the unique stage given
}
\description{
A function to find which stages have merged with a given unique stage from the merging details of a CEG.
}
