% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_trees_functions.R
\name{model_combiner}
\alias{model_combiner}
\title{Model Combiner}
\usage{
model_combiner(
  data,
  mod.background,
  mod.response,
  background.order = NULL,
  back.prior.input = NULL,
  resp.prior.input = NULL
)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if included) are the last two columns.}

\item{mod.background}{A StagedTree object modelling the background process.}

\item{mod.response}{A StagedTree object modelling the response process.}

\item{background.order}{An integer vector detailing how the default backgrounds from the [\code{\link[=background_extractor]{background_extractor()}}] function have been reordered in the augmented dataset. If integer $m$ is in position $n$ of the vector, than the background $m$ from [\code{\link[=background_extractor]{background_extractor()}}] is now background $n$ in the augmented data set.}

\item{back.prior.input}{A list where each element is a matrix corresponding to a Dirichlet prior distribution for each level of the background tree before model selection. If NULL, the prior is extracted from the background model and assumed equally distributed.}

\item{resp.prior.input}{A list where each element is a matrix corresponding to a Dirichlet prior distribution for each level of the response tree before model selection. If NULL, the prior is extracted from the response model and assumed equally distributed.}
}
\value{
A StagedTree object which is equivalent to combining mod.background and mod.response.
}
\description{
Takes a background model and response model and combines them in one full model.
}
\details{
Intended for use with Intermediate CEGs, this creates a StagedTree object which is equivalent to combining the background process in mod.background, with the response process in mod. response. This combined model can be used to compare to other methods of model selection, such as the score.
}
\examples{
mod1<-pceg(knee_pain_obs,2,TRUE,TRUE)
mod.back<-pceg(knee_pain_obs[,1:3],2,FALSE,FALSE)
newdata<-background_extractor(knee_pain_obs,mod.back)
mod.resp<-pceg(newdata,2,TRUE,TRUE)
mod2<-model_combiner(knee_pain_obs,mod.back,mod.resp)
}
