% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_optimisation_functions.R
\name{nlm_zip}
\alias{nlm_zip}
\title{The Numerical Optimisation Method}
\usage{
nlm_zip(data, variable_time = TRUE)
}
\arguments{
\item{data}{A data set, where the observed count vector and time vector (if variable) are the last two columns.}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE).}
}
\value{
A list containing a matrix with the desired outputs for each evolution of the process, as well as vectors of the estimates for the rates and risk probabilities.
}
\description{
The Numerical Optimisation Method
}
