% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{PairwisePosition1}
\alias{PairwisePosition1}
\title{The Pairwise Position Function}
\usage{
PairwisePosition1(pair.situation, num.category, pos.next.level)
}
\arguments{
\item{pair.situation}{An integer vector}

\item{num.category}{An integer value}

\item{pos.next.level}{An integer vector}
}
\value{
A logical value indicating whether two situations are in the same position.
}
\description{
The function \code{\link[ceg:PairWisePosition]{ceg::PairWisePosition()}}
}
