% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_functons.R
\name{path_refactor}
\alias{path_refactor}
\title{The Path refactor function}
\usage{
path_refactor(data, n, data_levels)
}
\arguments{
\item{data}{A data set where the covariates are factors.}

\item{n}{An integer for the number of columns to refactor, beginning from the start.}

\item{data_levels}{An integer vector of the number of levels for each covariate.}
}
\value{
A data set where the categorical covariates have been refactor to start at 0 and end at the number of levels - 1.
}
\description{
This function takes the categorical covariates from a data set and refactors each one from 0 to the number of levels - 1.
}
