% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions_S3.R
\name{plot.StagedTree}
\alias{plot.StagedTree}
\title{The \code{\link[=plot]{plot()}} method for the S3 Class StagedTree}
\usage{
plot.StagedTree(x)
}
\arguments{
\item{x}{An object in the S3 class StagedTree}
}
\value{
A plot of the StagedTree
}
\description{
The \code{\link[=plot]{plot()}} method for the S3 Class StagedTree
}
\examples{
mod<-pceg(knee_pain_obs,2,TRUE,TRUE)
plot(mod)
}
