% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{PositionLevel1}
\alias{PositionLevel1}
\title{The Position Level Function}
\usage{
PositionLevel1(
  stage.list,
  num.category,
  num.situation.next,
  pos.next.level = list()
)
}
\arguments{
\item{stage.list}{A list detailing the stage structure for a single level.}

\item{num.category}{An integer value detailing the number of stages at a single level.}

\item{num.situation.next}{An integer value detailing the number of situations at the next level.}

\item{pos.next.level}{A list detailing the position structure for the next level.}
}
\value{
A list detailing the position structure for a single level.
}
\description{
The function \code{\link[ceg:PositionLevel]{ceg::PositionLevel()}}
}
