% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{PositionStage1}
\alias{PositionStage1}
\title{The Position Stage Function}
\usage{
PositionStage1(stage.vector, num.category, pos.next.level)
}
\arguments{
\item{pos.next.level}{}
}
\value{

}
\description{
The Position Stage Function
}
