% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{PositionVector1}
\alias{PositionVector1}
\title{The Position Vector Function}
\usage{
PositionVector1(num.situation, pos.list)
}
\arguments{
\item{pos.list}{}
}
\value{

}
\description{
The Position Vector Function
}
