% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_new_functions.R
\name{stage_structure}
\alias{stage_structure}
\title{The Stage Structure Function}
\usage{
stage_structure(mod, score = 0)
}
\arguments{
\item{mod}{A list detailing the output from \code{\link[=pceg]{pceg()}}}

\item{score}{A numeric vector detailing the change in score for each level of the tree}
}
\value{
A list specifying the stage structure from the output of \code{\link[=pceg]{pceg()}} in a way that is compatible with the "Stratified.staged.tree" S4 class.
}
\description{
This function takes a pceg model and creates a stage structure that is compatible with the "Stratified.staged.tree" S4 class.
}
\examples{
mod1<-pceg(knee_pain_obs,2,TRUE,TRUE)
stage_structure(mod1)
}
