% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_package_updated_functions.R
\name{StratifiedCegPosition1}
\alias{StratifiedCegPosition1}
\title{The StratifiedCEGPosition1 Function}
\usage{
StratifiedCegPosition1(stage, num.category, num.situation)
}
\arguments{
\item{num.situation}{}
}
\value{

}
\description{
The StratifiedCEGPosition1 Function
}
