% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_extractor_function.R
\name{value_extractor}
\alias{value_extractor}
\title{The Expected Value Extractor}
\usage{
value_extractor(
  data,
  ceg,
  level_rel_final = 0,
  poisson_response = TRUE,
  variable_time = TRUE,
  posterior = TRUE,
  zip = TRUE,
  true_value = NA
)
}
\arguments{
\item{data}{A data set, where the observed response vector and time vector (if applicable and variable) are the last two columns}

\item{ceg}{A ceg model fit to the data set, as produced by pceg().}

\item{level_rel_final}{A non-positive integer indicating where the desired variable is relative to the response variable.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{posterior}{A logical value indicating whether the estimates of the posterior (TRUE) or sample (FALSE) expected value should be calculated.}

\item{zip}{A logical value indicating whether the model specified is zero-inflated (TRUE) or not (FALSE).}

\item{true_value}{A numeric vector specifying the true values of the parameters for comparison, if known. If unknown, this will be NA.}
}
\value{
A matrix displaying the observed data, the prior values, and the expected values of the parameters.
}
\description{
This function calculates the expected values of parameters based on the data set and a chosen pceg model.
}
