% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_functions.R
\name{zipceg}
\alias{zipceg}
\title{The ZIPCEG function}
\usage{
zipceg(
  data,
  method = "Gibbs",
  iter = 10000,
  equivsize = 2,
  poisson_response = TRUE,
  variable_time = FALSE,
  stoch_imputation = TRUE,
  gamma_alpha = 1,
  gamma_beta = 2,
  beta_c = 1,
  beta_d = 1,
  p_0 = NA,
  l_0 = NA,
  tol = 1e-10,
  var_disc = 0,
  disc_length = 0,
  restrict = FALSE,
  mirror = FALSE,
  cat_limit = 0
)
}
\arguments{
\item{data}{A data set where the observed count vector and time vector (if variable) are the last two columns.}

\item{method}{A character string indicating the method for parameter estimation. The character string can be an element of c("Gibbs","nlm","EM","mle","mm").}

\item{iter}{The number of iterations for the Gibbs sampler or Expectation-Maximisation algorithm.}

\item{equivsize}{A numeric value specifying the equivalent sample size for the prior, a measure of confidence in the prior.}

\item{poisson_response}{A logical value indicating whether the response variable is Poisson (TRUE) or categorical (FALSE).}

\item{variable_time}{A logical value indicating whether the observed time is uniform (FALSE) or variable (TRUE), if applicable.}

\item{stoch_imputation}{A logical value indicating whether stochastic (TRUE) or deterministic (FALSE) imputation should be used.}

\item{gamma_alpha}{A numeric value for the shape hyperparameter of the Gamma prior for the Poisson rate, if applicable.}

\item{gamma_beta}{A numeric value for the rate hyperparameter of the Gamma prior for the Poisson rate.}

\item{p_0}{A numeric initial value for the risk probability.}

\item{l_0}{A numeric initial value for the rate.}

\item{tol}{A numeric which represents the minimum change in the complete data log likelihood needed to continue the Expectation-Maximisation algorithm.}

\item{var_disc}{An integer value specifying which variable to discretise. If 0, no discretisation is necessary.}

\item{disc_length}{An integer value specifying how many neighbours can be searched over for the purposes of variable discetisation. If 0, all other possible stages may be merged over.}

\item{restrict}{A logical value indicating whether variable discretisation should be restricted to stages with the same unfolding of the process (TRUE) or not (FALSE).}

\item{mirror}{A logical value indicating whether variable discretisation should be equivalent across each unfolding of the process (TRUE) or not (FALSE).}

\item{cat_limit}{An integer value specifying the minimum number of categories to the variable can be discretised to. If 0, there is no minimum number of categories.}

\item{c}{A numeric value for the alpha hyperparameter of the Beta prior for the risk probability.}

\item{d}{A numeric value for the beta hyperparameter of the Beta prior for the risk probability.}
}
\value{
A list specifying a ZIPCEG model. The list contains: the prior for the final model, the data for the final model, the stage numbers for the final model,
the stage structure for the final model, the vector of likelihoods after each merging, the details of the stages merged at each step,
the comparison set of stages left to be merged, a list of the stages merged, and the log marginal likelihood of the final model.
}
\description{
This function fits a ZIPCEG model to the chosen data set, based on chosen methods of parameter estimation and state imputation.
}
\examples{
 zipceg(knee_pain_obs,"nlm",variable_time=TRUE)
}
